+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Modules that explicitly import something from the `ABS.StdLib` standard library will not implicitly import everything from `ABS.StdLib` anymore.  The behavior of modules not having such an import clause is unchanged, i.e., they still contain an impolicit `import * from ABS.StdLib` clause.
 
+- In a module definition, `import` and `export` clauses can now be written in any order.
+
 - Avoid crash in `DC.decrementResources`: when trying to decrement by more than available, only decrement to the maximum possible.  `decrementResources` now returns the actual amount by which the resource was decreased.  This changes the signature of three methods in the `ABS.DC.DeploymentComponent` interface in a backwards-incompatible way.
 
 - When pretty-printing to a file, the resulting file will use Unix line endings.